Did some cleaning up
Stopped the asteroid rotation animation moving at random speeds causing it to rotate weirdly sometimes. Fixed movement being able to be stopped by pressing any button. Tidied up some code and removed some which was not needed

+++ HelloGL/HelloGL/PowerUp.h
@@ -18,7 +18,7 @@ private:
 	Vector3 _position;
 	GLfloat _objectAcceleration;
 	bool slowPowerUpActive;
-	int slowPowerUpTimer;
+	float slowPowerUpTimer;
 
 public:
 	PowerUp(Mesh* mesh, Texture2D* texture, float x, float y, float z);
